==> train/notebooks/1.data_processing.ipynb <==
## Exploratory analysis of the tweet sentiment expression dataset.

### Introduction :  


The main objective of our project is to develop an intelligent system capable of analyzing and understanding the sentiment of a message posted by a user on a social network, in this case here [Twitter](https://twitter. com/home?lang=fr). Using advanced natural language processing (NLP) techniques, our solution will seek to extract key information, establish semantic connections and create a representation of user sentiment.

## Cleaning data

## Import

We will use a usual Data Science stack: `numpy`, `pandas`, `sklearn`, `matplotlib`.

# data manipulation
import numpy as np
import pandas as pd

# graphic representation
import matplotlib.pyplot as plt
import seaborn as sns

# file system management
import os

The data that we will use during our study has the following relative path ["../Data"](). We can find in the [./Data]() folder a csv file which includes a list of user tweets. It is in this folder that we will save our dataframes to be able to use them in other operations of our study.

print(os.listdir("../Data"))

We import the file which will be useful for our study.

data = pd.read_csv('../Data/tweets.csv', encoding='latin-1', header=None)
data = data.rename(columns={data.columns[0]: 'target'})
data = data.rename(columns={data.columns[1]: 'id'})
data = data.rename(columns={data.columns[2]: 'date'})
data = data.rename(columns={data.columns[3]: 'flag'})
data = data.rename(columns={data.columns[4]: 'user'})
data = data.rename(columns={data.columns[5]: 'text'})
data

The file contains 1,600,000 user tweets extracted using Twitter's API. The tweets have been annotated (0 = negative, 4 = positive) and can be used to detect the sentiment of a tweet.

The DataFrame is characterized by 6 variables:

1. target: the sentiment of the tweet (0 = negative, 4 = positive)
2. id: tweet id (2087)
3. date: tweet date (Sat May 16 23:58:44 UTC 2009)
4. flag: The query (LyX). If there is no query then the value will be NO_QUERY
5. user: the user who tweeted (robotickilldozr)
6. text: the text of the tweet (with LyX)

with pd.option_context('display.max_colwidth', None):
    print(data.iloc[0, :])

We apply text simplification by removing user identification, web links, single characters, numeric characters, non-alphanumeric characters.

import re
data['text'] = data['text'].apply(lambda x: re.sub(r'\S*@\S*\s?', '', x))
data['text'] = data['text'].apply(lambda x: re.sub(r'http\S+', '', x))
data['text'] = data['text'].apply(lambda x: re.sub(r'[^\w\s]', '', x))
data['text'] = data['text'].apply(lambda x: re.sub(r'\b\w\b', '', x))
data['text'] = data['text'].apply(lambda x: re.sub(r'\d', '', x))
data['text'] = data['text'].apply(lambda x: re.sub(r'\s+', ' ', x))
data['text'] = data['text'].apply(lambda x: x.lower())


with pd.option_context('display.max_colwidth', None):
    print(data.iloc[0, :])

The nltk library allows us to remove stopwords. These are words that recur in a language but do not provide additional information for understanding the text.

import nltk
from nltk.corpus import stopwords

nltk.download('stopwords')

data['text_token'] = 0

for i in range(len(data['text'])):
    data['text_token'][i] = nltk.word_tokenize(data['text'][i])

def remove_stopwords(word_list):
    stop_words = set(stopwords.words('english'))
    filtered_words = []
    for word in word_list:
        if word not in stop_words:
            filtered_words.append(word)
    return filtered_words

for i in range(len(data['text_token'])):
    data['text_token'][i] = remove_stopwords(data['text_token'][i])

with pd.option_context('display.max_colwidth', None):
    print(data.iloc[0, :])

import nltk
from nltk.stem import WordNetLemmatizer
from nltk.corpus import wordnet

nltk.download('punkt')# segmentation phrases
nltk.download('averaged_perceptron_tagger') # étiquettes grammaticales
nltk.download('wordnet')# synonymes

Lemmatization allows us to simplify the meaning of sentences by finding the root of each word. We thus remove the conjugation or even the plural from each word.

def lemmatize_words(word_list):
    lemmatizer = WordNetLemmatizer()
    lemmatized_words = [lemmatizer.lemmatize(word, get_pos(word)) for word in word_list]
    return lemmatized_words

def get_pos(word):
    tag = nltk.pos_tag([word])[0][1][0].upper()
    tag_dict = {"N": wordnet.NOUN, "V": wordnet.VERB, "R": wordnet.ADV, "J": wordnet.ADJ}
    return tag_dict.get(tag, wordnet.NOUN)

for i in range(len(data['text_token'])):
    data['text_token'][i] = lemmatize_words(data['text_token'][i])

word_dataset = '../Data/normalized_dataset.csv'
data.to_csv(word_dataset, index=False)

df = pd.read_csv("../Data/normalized_dataset.csv")

import ast
import re

array_list = df['text_token'].values
data_list = []
for item in array_list:
    data_list.append(ast.literal_eval(item))

df_list = pd.DataFrame({'text_token': data_list})
df = df.drop(columns=['text_token'])
df['text_token'] = df_list['text_token']
df['words'] = df['text_token'].apply(lambda x: ' '.join(x))
df = df.drop(columns=['text_token'])

with pd.option_context('display.max_colwidth', None):
    print(df.iloc[0, :])

word_dataset = '../Data/cleaned_dataset.csv'
df.to_csv(word_dataset, index=False)

df = pd.read_csv("../Data/cleaned_dataset.csv")

df_neg = df[df['target']== 0].sample(500000)
df_pos = df[df['target']== 4].sample(500000)
df_pos['target'] = 1
liste_concat = [df_neg, df_pos]
df_sample = pd.concat([df_neg, df_pos], ignore_index=True)
df_sample = df_sample.sample(frac=1).reset_index(drop=True)
df_sample

sample_df = '../Data/sample_dataset.csv'
df_sample.to_csv(sample_df, index=False)
