==> train/notebooks/2.train_test_split.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv("../Data/sample_dataset.csv")

from sklearn.model_selection import train_test_split

train_df, test_df = train_test_split(df, test_size=0.2, random_state=42)

saving_path = '../Data/train_df.csv'
train_df.to_csv(saving_path, index=False)

saving_path = '../Data/test_df.csv'
test_df.to_csv(saving_path, index=False)
