==> train/notebooks/6.mlflow.ipynb <==
import torch
import matplotlib.pyplot as plt
from transformers import RobertaTokenizer
from transformers import RobertaForSequenceClassification
from transformers import RobertaModel

tokenizer_path = "./tokenizer_roberta"
model_path = "./model_roberta"

tokenizer = RobertaTokenizer.from_pretrained(tokenizer_path)
model = RobertaForSequenceClassification.from_pretrained(model_path)

def predict_sentiment(text):
    inputs = tokenizer(text, return_tensors="pt", truncation=True, padding=True)
    

    with torch.no_grad():
        outputs = model(**inputs)

    logits = outputs.logits

    probabilities = torch.softmax(logits, dim=1).squeeze()

    predicted_class = torch.argmax(probabilities).item()
    predicted_sentiment = "positive" if predicted_class == 1 else "negative"

    return predicted_sentiment, probabilities


text = "Hello, how are you?"
sentiment, probabilities = predict_sentiment(text)
print("Sentiment prédit:", sentiment)
print("Probabilités:", probabilities)

tokenizer_path = "./tokenizer_roberta"
model_path = "./model_roberta"

tokenizer = RobertaTokenizer.from_pretrained(tokenizer_path)
model = RobertaModel.from_pretrained(model_path)

def interpret_sentence(sentence):
    inputs = tokenizer(sentence, return_tensors="pt", truncation=True, padding=True)
    input_ids = inputs["input_ids"]
    attention_mask = inputs["attention_mask"]

    # Calculer les embeddings
    with torch.no_grad():
        output = model(input_ids, attention_mask=attention_mask)

    logits = output.logits
    embeddings = model.roberta.embeddings.word_embeddings(input_ids)

    # Récupérer les couches d'attention
    attentions = output.attentions

    return embeddings, attentions

sentence = "Hello, how are you?"
embeddings, attentions = interpret_sentence(sentence)

import torch
from transformers import RobertaTokenizer, RobertaModel
import matplotlib.pyplot as plt
import seaborn as sns

tokenizer_path = "./tokenizer_roberta"
model_path = "./model_roberta"

tokenizer = RobertaTokenizer.from_pretrained(tokenizer_path)
model = RobertaModel.from_pretrained(model_path)
model.eval()

def interpret_sentence(sentence):
    inputs = tokenizer(sentence, return_tensors="pt", truncation=True, padding=True)
    input_ids = inputs["input_ids"]
    attention_mask = inputs["attention_mask"]

    # Calculer les embeddings
    with torch.no_grad():
        output = model(input_ids, attention_mask=attention_mask)

    embeddings = output.last_hidden_state
    attentions = output.attentions

    return embeddings, attentions

def visualize_attention(sentence, attentions):
    # Récupérer la longueur de la phrase
    n_tokens = len(tokenizer(sentence)['input_ids'])

    # Créer une heatmap pour chaque couche d'attention
    for layer, layer_attention in enumerate(attentions):
        plt.figure(figsize=(10, 6))
        sns.heatmap(layer_attention[0][:n_tokens, :n_tokens], cmap='viridis', annot=False)
        plt.title(f'Layer {layer+1} Attention')
        plt.xlabel('To')
        plt.ylabel('From')
        plt.show()

# Utiliser la fonction pour visualiser l'attention
sentence = "Hello, how are you?"
embeddings, attentions = interpret_sentence(sentence)
visualize_attention(sentence, attentions)


import pandas as pd

data = pd.read_csv('../Data/tweets.csv', encoding='latin-1', header=None)
data = data.rename(columns={data.columns[0]: 'target'})
data = data.rename(columns={data.columns[1]: 'id'})
data = data.rename(columns={data.columns[2]: 'date'})
data = data.rename(columns={data.columns[3]: 'flag'})
data = data.rename(columns={data.columns[4]: 'user'})
data = data.rename(columns={data.columns[5]: 'text'})
data = data.sample(1000)
data

from transformers import pipeline

classifier = pipeline("text-classification", model=model, tokenizer=tokenizer)
text_list = data['text'].tolist()
y_pred = classifier(text_list)

y_pred_tab = []
for i in y_pred:
    if i['label'] == 'LABEL_1':
        y_pred_tab.append(1)
    else :
        y_pred_tab.append(0)

y_test_tab = []
for i in range(len(data)):
    if data['target'].iloc[i] == 0:
        y_test_tab.append(0)
    else :
        y_test_tab.append(1)

from sklearn.metrics import accuracy_score

base_accuracy = accuracy_score(y_test_tab, y_pred_tab)

base_accuracy

def interpret_sentence(sentence):
    inputs = tokenizer(sentence, return_tensors="pt", truncation=True, padding=True)
    input_ids = inputs["input_ids"]
    attention_mask = inputs["attention_mask"]

    # Calculer les embeddings
    with torch.no_grad():
        output = model(input_ids, attention_mask=attention_mask)

    logits = output.logits
    embeddings = model.roberta.embeddings.word_embeddings(input_ids)

    attentions = output.attentions

    return embeddings, attentions


sentence = "Go fuck yourself i hate you"
embeddings, attentions = interpret_sentence(sentence)

from captum.attr import LayerIntegratedGradients, TokenReferenceBase, visualization

visualization.visualize_text_attention(sentence, attentions)

import seaborn as sns
import matplotlib.pyplot as plt

def visualize_attention(sentence, attentions):
    # Récupérer la longueur de la phrase
    n_tokens = len(tokenizer(sentence)['input_ids'])

    # Créer une heatmap pour chaque couche d'attention
    for layer, layer_attention in enumerate(attentions):
        plt.figure(figsize=(10, 6))
        sns.heatmap(layer_attention[0][:n_tokens, :n_tokens], cmap='viridis', annot=False)
        plt.title(f'Layer {layer+1} Attention')
        plt.xlabel('To')
        plt.ylabel('From')
        plt.show()

# Utiliser la fonction pour visualiser l'attention
visualize_attention(sentence, attentions)
